laboratory_work_3/weather/serializers.py
from rest_framework import serializers

from weather.models import User, Town, Country


class UserCreateSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['email', 'username', 'password', 'first_name', 'last_name', 'days_count', 'towns']


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['email', 'username', 'first_name', 'last_name', 'days_count', 'towns']


class TownSerializer(serializers.ModelSerializer):
    class Meta:
        model = Town
        fields = '__all__'


class CountrySerializer(serializers.ModelSerializer):
    class Meta:
        model = Country
        fields = '__all__'
