laboratory_work_3/requirements.txt
Django==3.2
djangorestframework
djangorestframework-simplejwt
djoser
drf-spectacular
